Split out primitive parsers into separate file
